+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Display chronostratigraphy colors in chronostratigraphy editor.
 - Migrate chronostratigraphy _Middle Jurassic undifferenciated_ to _Middle Jurassic_.
 - Remove height reference system column from editor view list.
+- Remove import settings from administrator settings.
 
 ### Fixed